+++ ApocalypseSnow/ShotStruct.cs
@@ -3,8 +3,8 @@
 public struct ShotStruct
 {
     public MessageType Type;
-    public int mouseX;
-    public int mouseY;
+    public float mouseX;
+    public float mouseY;
     public int charge; // <-- Nuova variabile aggiunta
     
     public ShotStruct()
